pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid-ssl"
version = "0.1.0"
description = "Hybrid MAE + SimCLR self-supervised learning"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Hybrid SSL Team"}]
license = {text = "MIT"}
dependencies = [
    "torch==2.2.2",
    "torchvision==0.17.2",
    "torchaudio==2.2.2",
    "numpy==1.26.4",
    "scipy==1.11.4",
    "scikit-learn==1.4.1",
    "matplotlib==3.8.3",
    "seaborn==0.13.2",
    "tqdm==4.66.2",
    "pyyaml==6.0.1",
    "tensorboard==2.15.2",
    "wandb==0.16.4",
    "timm==0.9.12",
    "einops==0.7.0",
    "umap-learn==0.5.5",
    "pandas==2.2.1",
    "huggingface-hub==0.20.2",
    "datasets==2.17.1",
]

[project.optional-dependencies]
develop = ["pytest", "black", "ruff"]

[tool.setuptools.packages.find]
exclude = ["tests", "docs"]
